add checkbox to control sql auto-formatting in editor
