+++ public/components/tables/results.js
@@ -20,6 +20,15 @@ export function initQueryExecution() {
     window.sqlEditor.on('change', handleEditorChange);
 
     /**
+    * Checks if SQL formatting in editor is enabled
+    * @returns {boolean} True if formatting should be applied to the editor
+    */
+    function shouldFormatInEditor() {
+        const formatSqlCheckbox = document.getElementById('format-sql-checkbox');
+        return formatSqlCheckbox ? formatSqlCheckbox.checked : true;
+    }
+
+    /**
      * Set the behavior to trigger when the SQL editor content changes.
      * 
      * Depending on whether this content is empty or not, it will:
@@ -69,7 +78,10 @@ export function initQueryExecution() {
             functionCase: 'lower',
             linesBetweenQueries: 2,
         });
-        window.sqlEditor.setValue(query);
+        if (shouldFormatInEditor()) {
+            window.sqlEditor.setValue(query);
+        }
+
         const result = await runQueryAndRenderResults(query);
         executionTab.removeEventListener('click', triggerExecutionOfNewQuery);
         setTabIconTo(checkIcon);
@@ -100,7 +112,7 @@ export function initQueryExecution() {
             // Render the paginated table with results and page change handler
             renderPaginatedTable(data, resultsContainer, changePage);
             return data;
-            
+
         } catch (error) {
             showError(error.message, resultsContainer);
             return null;
