+++ public/index.html
@@ -53,7 +53,16 @@
                         </svg>
                     </div>
                 </div>
+                <!-- SQL Formatting option -->
+                <div class="menu-item formatting-container">
+                    <span class="menu-label" data-i18n="header.formatSql">Format SQL in editor</span>
+                    <div class="checkbox-container">
+                        <input type="checkbox" id="format-sql-checkbox" checked>
+                        <label for="format-sql-checkbox" class="checkbox-label"></label>
+                    </div>
+                </div>
             </div>
+        </div>
         </div>
 
         <!-- Title -->
